move forehead stuff to its own file
