+++ webserver/static/js/feverscreen.ts
@@ -1,18 +1,17 @@
 import {BlobReader} from "./utils.js";
-import {fahrenheitToCelsius, moduleTemperatureAnomaly, sensorAnomaly, ROIFeature, FeatureState, Delta} from "./processing.js";
+import {fahrenheitToCelsius, moduleTemperatureAnomaly, sensorAnomaly, ROIFeature} from "./processing.js";
 import {DeviceApi} from "./api.js";
 import {CalibrationInfo, FrameInfo, Modes, NetworkInterface, TemperatureSource} from "./feverscreen-types.js";
 import {circleDetect, circleDetectRadius, edgeDetect} from "./circledetect.js";
 import {buildSAT, scanHaar, ConvertCascadeXML, HaarCascade} from "./haarcascade.js";
+import {detectForehead} from "./forehead-detect.js";
 
 let GForeheads: ROIFeature[];
 let GROI: ROIFeature[] = [];
 
-// top percent to be considered forehead
-const ForeheadPercent = 0.3;
-const ForeheadPadding = 2;
+
 const ForeheadColour = "#00ff00";
-const ForeheadEdgeThresh = 200;
+
 const GSensor_response = 0.030117;
 const GDevice_sensor_temperature_response = -30.0;
 
@@ -1199,288 +1198,6 @@ window.onload = async function() {
     return x * alpha + y * (1 - alpha);
   }
 
-  //  uses the sobel operator, to return the intensity and direction of edge at
-  // index
-  function sobelEdge(
-    source: Float32Array,
-    index: number,
-    width: number
-  ): [number, number] {
-    const x = sobelX(source, index, width);
-    const y = sobelY(source, index, width);
-
-    return [Math.sqrt(x * x + y * y), Math.atan(y / x)];
-  }
-
-  function sobelY(source: Float32Array, index: number, width: number): number {
-    return (
-      -source[index - 1 - width] -
-      2 * source[index - width] -
-      source[index - width + 1] +
-      source[index - 1 + width] +
-      2 * source[index + width] +
-      source[index + width + 1]
-    );
-  }
-
-  function sobelX(source: Float32Array, index: number, width: number): number {
-    return (
-      -source[index - 1 - width] +
-      source[index + 1 - width] -
-      2 * source[index - 1] +
-      2 * source[index + 1] -
-      source[index - 1 + width] +
-      source[index + 1 + width]
-    );
-  }
-
-  function featureLine(x: number, y: number): ROIFeature {
-    let line = new ROIFeature();
-    line.y0 = y;
-    line.y1 = y;
-    line.x0 = x;
-    line.x1 = x;
-    line.state = FeatureState.Outside;
-    return line;
-  }
-
-  // scan left to right of the range of faceX
-  // at each x detect edges of the face with a vertical line
-  // returns the longest y line
-  function yScan(
-    source: Float32Array,
-    faceX: ROIFeature,
-    roi: ROIFeature,
-    endFaceY: number | null
-  ): ROIFeature | undefined {
-    let endY;
-    let midY;
-
-    if (endFaceY) {
-      midY = ~~((roi.y0 + endFaceY) / 2);
-      endY = Math.min(frameHeight - 1, endFaceY);
-    } else {
-      midY = ~~roi.midY();
-      endY = Math.min(frameHeight - 1, ~~roi.y1 - 1);
-    }
-    let longestLine;
-
-    for (let x = ~~faceX.x0 + 1; x < ~~faceX.x1 - 1; x++) {
-      let faceY = featureLine(x, -1);
-      for (let y = ~~roi.y0 + 1; y < endY; y++) {
-        let index = y * frameWidth + x;
-        let [intensity, direction] = sobelEdge(source, index, frameWidth);
-        if (y == midY) {
-          // end any previous state
-          nextYState(faceY, false);
-        }
-        let edge = detectYEdge(faceY, y, midY, intensity, direction);
-        nextYState(faceY, edge);
-      }
-
-      if (faceY.hasYValues() && faceY.higher(longestLine)) {
-        longestLine = faceY;
-      }
-    }
-    return longestLine;
-  }
-
-  // for the line defined by faceY
-  // check if the edge  values (intensity, direction) justify an edge
-  function detectYEdge(
-    faceY: ROIFeature,
-    y: number,
-    midY: number,
-    intensity: number,
-    direction: number
-  ) {
-    if (intensity - ForeheadEdgeThresh < 0) {
-      return false;
-    }
-
-    if (y < midY) {
-      if (
-        faceY.state == FeatureState.Inside ||
-        (direction > -Math.PI / 4 - 0.5 && direction < Math.PI / 4 - 0.5)
-      ) {
-        return false;
-      }
-
-      //get strongest gradient of topmost edge
-      if (faceY.y0 == -1) {
-        faceY.y0 = y;
-        faceY.sensorY = intensity;
-      } else if (faceY.onEdge() && intensity > faceY.sensorY) {
-        faceY.sensorY = intensity;
-        faceY.y0 = y;
-      }
-    } else {
-      if (direction < -Math.PI / 4 - 0.5 || direction > Math.PI / 4 + 0.5) {
-        return false;
-      }
-
-      //get strongest gradient of bottommost edge
-      if (faceY.onEdge()) {
-        if (intensity > faceY.sensorY) {
-          faceY.sensorY = intensity;
-          faceY.y1 = y;
-        }
-      } else {
-        faceY.y1 = y;
-        faceY.sensorY = intensity;
-      }
-    }
-    return true;
-  }
-
-  function nextXState(r: ROIFeature, edge: boolean) {
-    if (r.state == FeatureState.Outside && edge) {
-      r.state = FeatureState.LeftEdge;
-    } else if (r.state == FeatureState.LeftEdge && !edge) {
-      r.state = FeatureState.Inside;
-    } else if (r.state == FeatureState.Inside && edge) {
-      r.state = FeatureState.RightEdge;
-    } else if (r.state == FeatureState.RightEdge && !edge) {
-      r.state = FeatureState.Outside;
-    } else if (r.state == FeatureState.Outside && edge) {
-      r.state = FeatureState.RightEdge;
-    }
-  }
-
-  function nextYState(r: ROIFeature, edge: boolean) {
-    if (r.state == FeatureState.Outside && edge) {
-      r.state = FeatureState.TopEdge;
-    } else if (r.state == FeatureState.TopEdge && !edge) {
-      r.state = FeatureState.Inside;
-    } else if (r.state == FeatureState.Inside && edge) {
-      r.state = FeatureState.BottomEdge;
-    } else if (r.state == FeatureState.BottomEdge && !edge) {
-      r.state = FeatureState.Outside;
-    } else if (r.state == FeatureState.Outside && edge) {
-      r.state = FeatureState.BottomEdge;
-    }
-  }
-
-  // for the line defined by faceY
-  // check if the edge  values (intensity, direction) justify an edge
-  function detectXEdge(
-    faceX: ROIFeature,
-    x: number,
-    midX: number,
-    intensity: number,
-    direction: number
-  ): boolean {
-    if (intensity - ForeheadEdgeThresh < 0) {
-      return false;
-    }
-    if (x < midX) {
-      if (
-        FeatureState.Inside == faceX.state ||
-        direction < -0.5 ||
-        direction > Math.PI / 2 + 0.5
-      ) {
-        return false;
-      }
-
-      // get strong gradient of left most edge
-      if (faceX.state == FeatureState.Outside) {
-        faceX.x0 = x;
-        faceX.sensorX = intensity;
-      } else if (faceX.onEdge() && intensity > faceX.sensorX) {
-        faceX.sensorX = intensity;
-        faceX.x0 = x;
-      }
-    } else {
-      if (direction > 0.5 && direction < Math.PI / 2 - 0.5) {
-        return false;
-      }
-
-      // get strongest gradient of right most edge
-      if (faceX.onEdge()) {
-        if (intensity > faceX.sensorX) {
-          faceX.sensorX = intensity;
-          faceX.x1 = x;
-        }
-      } else {
-        faceX.x1 = x;
-        faceX.sensorX = intensity;
-      }
-    }
-    return true;
-  }
-
-  // scan top to bottom of the range of faceY
-  // at each y detect edges of the face with a horizontal line
-  // attempts to find the end of the face by the width decreasing then increasing
-  // returns the longest horizontal line, and suspected end of the face y
-  function xScan(
-    source: Float32Array,
-    faceY: ROIFeature,
-    roi: ROIFeature
-  ): [ROIFeature | undefined, number | null] {
-    let longestLine;
-    let deltaIncreased = false;
-    let widthDelta = new Delta();
-    const midX = ~~roi.midX();
-    for (let y = ~~faceY.y0 + 1; y < ~~faceY.y1 - 1; y++) {
-      let faceX = featureLine(-1, y);
-      for (let x = ~~roi.x0 + 1; x < ~~roi.x1 - 1; x++) {
-        if (x == midX) {
-          // end any previous edges
-          nextXState(faceX, false);
-        }
-        let index = y * frameWidth + x;
-        let [intensity, diretion] = sobelEdge(source, index, frameWidth);
-        let edgeDetected = detectXEdge(faceX, x, midX, intensity, diretion);
-        nextXState(faceX, edgeDetected);
-      }
-
-      if (faceX.hasXValues()) {
-        widthDelta.add(faceX.width());
-
-        if (
-          deltaIncreased &&
-          widthDelta.state == 1 &&
-          widthDelta.prevState == -1
-        ) {
-          return [longestLine as ROIFeature, y];
-        } else if (widthDelta.state == 1) {
-          // gotta get bigger first
-          deltaIncreased = true;
-        }
-
-        if (faceX.wider(longestLine)) {
-          longestLine = faceX;
-        }
-      }
-    }
-    return [longestLine, null];
-  }
-
-  // scan the haar detected rectangle along y axis, to find range of x values,
-  // then along the x axis to find the range of y values
-  // choose the biggest x and y value to define xRad and yRad of the head
-  function detectForehead(
-    roi: ROIFeature,
-    source: Float32Array
-  ): ROIFeature | null {
-    let [faceX, endY] = xScan(source, roi, roi);
-    if (!faceX) {
-      faceX = roi;
-    }
-    let faceY = yScan(source, faceX, roi, endY);
-    if (!faceY) {
-      return null;
-    }
-
-    let forehead = new ROIFeature();
-    forehead.y0 = faceY.y0 - ForeheadPadding;
-    forehead.y1 = faceY.y0 + faceY.height() * ForeheadPercent + ForeheadPadding;
-    forehead.x0 = faceX.x0 - ForeheadPadding;
-    forehead.x1 = faceX.x1 + ForeheadPadding;
-    return forehead;
-  }
-
   function setSimpleHotSpot(r: ROIFeature, source: Float32Array) {
     for (let y = ~~r.y0; y < ~~r.y1; y++) {
       for (let x = ~~r.x0; x < ~~r.x1; x++) {
@@ -1520,7 +1237,7 @@ window.onload = async function() {
     GForeheads = [];
     for (let i = 0; i < roi.length; i++) {
       if (roi[i].flavor != "Circle") {
-        let forehead = detectForehead(roi[i], smoothedData);
+        let forehead = detectForehead(roi[i], smoothedData, frameWidth, frameHeight);
         if (forehead) {
           setSimpleHotSpot(forehead, smoothedData);
           roi[i].sensorX = forehead.sensorX;
